[MP-009] Refactoring: base class for common Kafka logic, stricter type guarantees for JobPaginator (using ChunkedJobMetadata).
