+++ scheduler-jobrunr/src/main/kotlin/io/scheduler/comparison/jobrunr/jobs/handlers/pagination/PaginatedJobRequestHandler.kt
@@ -2,10 +2,11 @@ package io.scheduler.comparison.jobrunr.jobs.handlers.pagination
 
 import io.scheduler.comparison.jobrunr.jobs.state.JobState
 import io.scheduler.comparison.jobrunr.jobs.pagination.JobPaginator
+import io.scheduler.comparison.jobrunr.jobs.state.data.ChunkedJobMetadata
 import org.jobrunr.jobs.lambdas.JobRequest
 import org.jobrunr.jobs.lambdas.JobRequestHandler
 
-interface PaginatedJobRequestHandler<T : JobState<*, *>, V, K : JobRequest> : JobRequestHandler<K> {
+interface PaginatedJobRequestHandler<T : JobState<*, ChunkedJobMetadata>, V, K : JobRequest> : JobRequestHandler<K> {
 
     fun handleNextPage(paginator: JobPaginator<T, V>): Collection<V>
