+++ scheduler-jobrunr/src/main/kotlin/io/scheduler/comparison/jobrunr/jobs/pagination/JobPaginator.kt
@@ -1,8 +1,9 @@
 package io.scheduler.comparison.jobrunr.jobs.pagination
 
 import io.scheduler.comparison.jobrunr.jobs.state.JobState
+import io.scheduler.comparison.jobrunr.jobs.state.data.ChunkedJobMetadata
 
-interface JobPaginator<T : JobState<*, *>, V> : Iterator<List<V>> {
+interface JobPaginator<T : JobState<*, ChunkedJobMetadata>, V> : Iterator<List<V>> {
 
     val jobState: T
     val pageSize: Int
